requirements.txt
torch==2.5.1
sentence-transformers==3.3.1
transformers==4.46.0
faiss-cpu==1.9.0
python-dotenv==1.0.1
python-telegram-bot==21.9
tqdm==4.67.1
llama_index==0.12.2
llama-index-embeddings-huggingface==0.4.0
llama-index-llms-huggingface==0.4.0
llama-index-vector-stores-faiss==0.3.0